+++ kernprof.py
@@ -13,7 +13,7 @@ from argparse import ArgumentError, ArgumentParser
 
 # NOTE: This version needs to be manually maintained in
 # line_profiler/line_profiler.py as well
-__version__ = '4.0.4'
+__version__ = '4.1.0'
 
 # Guard the import of cProfile such that 3.x people
 # without lsprof can still use this script.
@@ -239,6 +239,10 @@ def main(args=None):
         import line_profiler
         prof = line_profiler.LineProfiler()
         options.builtin = True
+        # Overwrite the explicit profile decorator
+        from line_profiler import explicit_profiler
+        explicit_profiler.__dict__['profile'] = prof
+        line_profiler.__dict__['profile'] = prof
     else:
         prof = ContextualProfile()
     if options.builtin: